pyproject.toml
[project]
name = "openrecall"
version = "0.8"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Flask==3.0.3",
    "numpy==1.26.4",
    "mss==9.0.1",
    "sentence-transformers==3.0.0",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "shapely==2.0.4",
    "h5py==3.11.0",
    "rapidfuzz==3.9.3",
    "Pillow==10.3.0",
    "python-doctr @ git+https://github.com/koenvaneijk/doctr.git@af711bc04eb8876a7189923fb51ec44481ee18cd",
]

[project.optional-dependencies]
windows = ["pywin32", "psutil"]
macos = ["pyobjc==10.3"]
linux = []
